Fix flakey localization UI test (#819)
