+++ change/@internal-react-composites-4a7f9c75-696c-402c-aa09-df45ea637bfa.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "Fix flakey ui tests",
+  "packageName": "@internal/react-composites",
+  "email": "2684369+JamesBurnside@users.noreply.github.com",
+  "dependentChangeType": "none"
+}
