+++ packages/react-composites/tests/browser/call/CallComposite.test.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-import { waitForCallCompositeToLoad, dataUiId, disableAnimation } from '../utils';
+import { waitForCallCompositeToLoad, dataUiId, disableAnimation, loadCallScreenWithParticipantVideos } from '../utils';
 import { test } from './fixture';
 import { expect, Page } from '@playwright/test';
 
@@ -41,8 +41,11 @@ test.describe('Call Composite E2E Tests', () => {
       expect(await page.screenshot()).toMatchSnapshot(`page-${idx}-local-device-settings-camera-disabled.png`);
       await page.click(dataUiId('call-composite-local-device-settings-microphone-button'));
       await page.click(dataUiId('call-composite-local-device-settings-camera-button'));
-      await page.waitForSelector('video');
-      await page.waitForTimeout(1000);
+      await page.waitForFunction(() => {
+        const videoNode = document.querySelector('video');
+        const videoLoaded = videoNode?.readyState === 4;
+        return !!videoNode && videoLoaded;
+      });
       expect(await page.screenshot()).toMatchSnapshot(`page-${idx}-local-device-settings-camera-enabled.png`);
     }
   });
@@ -53,25 +56,10 @@ test.describe('Call Composite E2E CallScreen Tests', () => {
   test.beforeEach(async ({ pages }) => {
     // In case it is retry logic
     for (const page of pages) {
-      page.reload();
-      await page.bringToFront();
+      await page.reload();
       await waitForCallCompositeToLoad(page);
-
-      await page.waitForSelector(dataUiId('call-composite-start-call-button'));
-      await page.click(dataUiId('call-composite-local-device-settings-camera-button'));
-      await page.click(dataUiId('call-composite-start-call-button'));
     }
-
-    // Wait for all video feeds to have loaded for all participants (local and remote)
-    for (const page of pages) {
-      await page.bringToFront();
-      await page.waitForFunction(() => {
-        const videoNodes = document.querySelectorAll('video');
-        const correctNoOfVideos = document.querySelectorAll('video').length === 2;
-        const allVideosLoaded = Array.from(videoNodes).every((videoNode) => videoNode.readyState === 4);
-        return correctNoOfVideos && allVideosLoaded;
-      });
-    }
+    await loadCallScreenWithParticipantVideos(pages);
   });
 
   test('video gallery renders for all pages', async ({ pages }) => {
